working towards sub commands
